+++ cpp/src/quote_context.cpp
@@ -1287,10 +1287,45 @@ QuoteContext::security_list(
   SecurityListCategory category,
   AsyncCallback<QuoteContext, std::vector<Security>> callback) const
 {
+  lb_security_list_category_t category_c = convert(category);
   lb_quote_context_security_list(
     ctx_,
     convert(market),
-    convert(category),
+    &category_c,
+    [](auto res) {
+      auto callback_ptr =
+        callback::get_async_callback<QuoteContext, std::vector<Security>>(
+          res->userdata);
+      QuoteContext ctx((const lb_quote_context_t*)res->ctx);
+      Status status(res->error);
+
+      if (status) {
+        auto rows = (const lb_security_t*)res->data;
+        std::vector<Security> rows2;
+        std::transform(rows,
+                       rows + res->length,
+                       std::back_inserter(rows2),
+                       [](auto row) { return convert(&row); });
+
+        (*callback_ptr)(AsyncResult<QuoteContext, std::vector<Security>>(
+          ctx, std::move(status), &rows2));
+      } else {
+        (*callback_ptr)(AsyncResult<QuoteContext, std::vector<Security>>(
+          ctx, std::move(status), nullptr));
+      }
+    },
+    new AsyncCallback<QuoteContext, std::vector<Security>>(callback));
+}
+
+void
+QuoteContext::security_list(
+  Market market,
+  AsyncCallback<QuoteContext, std::vector<Security>> callback) const
+{
+  lb_quote_context_security_list(
+    ctx_,
+    convert(market),
+    nullptr,
     [](auto res) {
       auto callback_ptr =
         callback::get_async_callback<QuoteContext, std::vector<Security>>(