+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [3.0.12] 2025-08-08
 
 - add `trade_session` for query all session intraday.
+- add `Market.Crypto`.
 
 # [3.0.10] 2025-07-27